fix: 저장하는 데이터 타입을 float16에서 float32로 변경

+++ train/preprocess.py
@@ -20,7 +20,7 @@ for i in range(0, len(dataset), 8192):
     for j, x in enumerate(dataset['sentence'][i:i+8192]):
         print(f"\rProcessing... {i+j+1:6d}/{len(dataset)}", end="")
         x = m.morphs(x)
-        x = np.array([model.wv[decomposeHangulText(word)] for word in x], dtype=np.float16)
+        x = np.array([model.wv[decomposeHangulText(word)] for word in x], dtype=np.float32)
         X[str(j)] = x
     np.savez_compressed(f"train/data/dataset/dataset_{i//8192:02d}_X.npz", **X)
